Inference refactor (#187)
