+++ src/b3d/chisight/gen3d/README.md
@@ -1,0 +1,3 @@
+# Gen3D
+
+You can run the `run_ycbv_evaluation.py` file as a script to evaluate Gen3D against the YCB-Video dataset.
